pyproject.toml
[tool.poetry]
name = "challenge"
version = "0.1.0"
description = ""
authors = ["asuraCoder <asuracoder@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1.3"
djangorestframework = "^3.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
